Load data from "backend"
This loads from a mock data file which will be replaced by an endpoint.
Uses fetch. Why fetch? It appears to be the future way to do AJAX as
well as the easiest way to get it working in my dev setup (just a single
npm install-save). Will need to test against various browsers to make
sure the polyfill is working correctly.